Simulation Configuration

UltraDark.Config — Module

The Config module contains structs and utility functions for controlling the simulation.

#Exports

SimulationConfig

TimeStepOptions

constant_scale_factor

UltraDark.Config.SimulationConfig — Type

struct SimulationConfig

Fields

a::Any: function defining scale factor a(t)

phase_grad_limit::Float64

density_threshold::Float64

time_step_options::Any

UltraDark.Config.TimeStepOptions — Type

struct TimeStepOptions

struct containing options controlling the size and calculation of time steps.

See also: SimulationConfig

Examples

julia> using UltraDark

julia> TimeStepOptions()
TimeStepOptions(10, 1.0)

Fields

update_period::Int64: controls how many steps are taken before the timestep is updated

multiplier::Float64: multiplies the calculated maximum time step by a constant

UltraDark.Config.TimeStepOptions — Method

TimeStepOptions(; update_period=10, multiplier=1.0)

UltraDark.Config.constant_scale_factor — Method

constant_scale_factor(t) -> Float64

A function that always returns 1.0, useful for simulations with no background expansion.