Summary statistics

It is often useful to collect summary statistics as a simulation runs and avoid storing complete simulation boxes. The UltraDark.Summary module makes this easy.

Example

Say you want to extract the minimum density at each time step. Define a struct to contain this data:

struct MinDensity
    ρx_min::Float64
end

and a function to construct it

function MaxDensity(sim_time, a, Δt, grids, constants, external_states)
    ρx_min = minimum(grids.ρx)

    MinDensity(ρx_max)
end

MaxDensity (generic function with 1 method)

Passing this object into a simulation will add a column to $output_path/summary.csv containing the single value contained by MinDensity.

More complex statistics are also possible.

Docstrings

UltraDark.Summary — Module

Summary

Module containing utilities for computing and outputing summary statistics at each time step.

UltraDark.Summary.EnergyGravity — Type

EnergyGravity

Gravitational potential energy

UltraDark.Summary.EnergyKineticQuantum — Type

EnergyKineticQuantum

Gravitational potential energy

UltraDark.Summary.MaxDensity — Type

MaxDensity

UltraDark.Summary.MaxDensityIndex — Type

MaxDensityIndex

This struct contains 4 useful pieces of information: the maxiumum value and three indices. Each is output to a summary.

UltraDark.Summary.MeanDensity — Type

MeanDensity

Summary statistic containing the mean density and the number of cells over which it was calculated.

Examples

julia> using UltraDark

julia> g = Grids(1.0, 16);

julia> Summary.MeanDensity(g)
UltraDark.Summary.MeanDensity(0.0, 4096)

UltraDark.Summary.RmsDensityContrast — Type

RmsDensityContrast

UltraDark.Summary.TotalMass — Type

TotalMass

Total mass on a grid

Examples

julia> using UltraDark

julia> g = Grids(1.0, 16);

julia> Summary.TotalMass(0.0, 1.0, 1e-1, g, nothing, ())
UltraDark.Summary.TotalMass(0.0)

UltraDark.Summary.WallTime — Type

WallTime

The current time in the real world.

Examples

WallTime created after another contains a later time.

julia> using UltraDark

julia> t1 = Summary.WallTime();

julia> t2 = Summary.WallTime(0.0, 1.0, 1e-1, Grids(1.0, 16), nothing, ());

julia> t1.date <= t2.date
true

UltraDark.Summary.column_title — Method

column_title(summary_struct)

Get column title

Assumes that the desired column title is the name of the first field. This can be refined for other structs by defining a specialization for the given datatype.

Note that the input argument is of type Val{T}, so that one can dispatch on T.

UltraDark.Summary.column_titles — Method

column_titles(summaries)

Generate column titles from an iterable of summaries

UltraDark.Summary.generate_summary_row — Method

generate_summary_row(summaries)

UltraDark.Summary.get_relevant_data — Method

get_relevant_data(summary_struct)

Format column entry as string

By default, assumes that the desired data is in the first field of the struct. This can be refined for other structs by defining a specialization for the given datatype and returning a comma separated string.

UltraDark.Summary.output_summary_header — Method

output_summary_header(output_config)

Write a header for a summary file

The header contains labels for each column of the summary CSV file. This function overwrites the current contents of the file.

UltraDark.Summary.output_summary_row — Method

output_summary_row(grids, output_config, t, a, Δt)

Write a new row to the summary file

UltraDark.Summary.pool_summarystat — Method

pool_summarystat(S1::MaxDensity, S2::MaxDensity)

MPI reduction operator for max density

UltraDark.Summary.pool_summarystat — Method

pool_summarystat(S1::MaxDensity, S2::MaxDensity)

MPI reduction operator for max density index

UltraDark.Summary.pool_summarystat — Method

pool_summarystat(S1::MeanDensity, S2::MeanDensity)

MPI reduction operator for mean density

UltraDark.Summary.pool_summarystat — Method

pool_summarystat(S1::RmsDensityContrast, S2::RmsDensityContrast)

MPI reduction operator for summary statistics.

UltraDark.Summary.pool_summarystat — Method

pool_summarystat(S1::ScaleFactor, S2::ScaleFactor)

MPI reduction operator for scale factor

Check that scale factors are equal and return

UltraDark.Summary.pool_summarystat — Method

pool_summarystat(S1::SimulationTime, S2::SimulationTime)

MPI reduction operator for simulation time

Check that times are equal and return

UltraDark.Summary.pool_summarystat — Method

pool_summarystat(S1::TimeStep, S2::TimeStep)

MPI reduction operator for scale factor

Check that scale factors are equal and return

UltraDark.Summary.pool_summarystat — Method

pool_summarystat(S1::WallTime, S2::WallTime)

MPI reduction operator for wall time

Return the time of the first argument.