UltraDark.jl

UltraDark.Grids

UltraDark.Grids

UltraDark.Grids

UltraDark.PencilGrids

UltraDark.PencilGrids

UltraDark.PencilGrids

UltraDark.E_grav

UltraDark.E_kq

UltraDark.actual_time_step

UltraDark.add_external_potential!

UltraDark.angular_momentum

UltraDark.angular_momentum_density

UltraDark.auxiliary_step!

UltraDark.azimuthal_angle

UltraDark.dV

UltraDark.evolve_to!

UltraDark.good_phase_diff

UltraDark.inner_step!

UltraDark.k_vec

UltraDark.mass

UltraDark.max_normed_phase_diff

UltraDark.max_time_step

UltraDark.max_time_step

UltraDark.outer_step!

UltraDark.phase_diff

UltraDark.phase_diff

UltraDark.polar_angle

UltraDark.radius_cylindrical

UltraDark.radius_spherical

UltraDark.take_steps!

UltraDark.Grids — Type

struct containing grids used in a simulation

Examples

julia> using UltraDark

julia> len = 1;

julia> resol = 16;

julia> Grids(len, resol);

UltraDark.Grids — Method

Grids(length, resol::Int)

Constructor for Grids

Create an empty grid with length length and resolution resol

Examples

julia> using UltraDark

julia> Grids(1.0, 64);

UltraDark.Grids — Method

Grids(length_tuple, resol_tuple::Tuple{Int, Int, Int})

Constructor for Grids

Create an empty length[1]xlength[2]xlength[3] grid with resolution resol[1]xresol[2]xresol[3].

Examples

julia> using UltraDark

julia> Grids((1.0, 1.0, 0.5), (64, 64, 32));

UltraDark.PencilGrids — Type

struct containing grids used in a simulation

Examples

julia> using UltraDark

julia> len = 1;

julia> resol = 16;

julia> PencilGrids(len, resol);

UltraDark.PencilGrids — Method

PencilGrids(length, resol)

Constructor for PencilGrids

Create an empty grid with length length and resolution resol. Uses PencilFFTs to create PencilArrays.

Examples

julia> using UltraDark

julia> PencilGrids(1.0, 64);

UltraDark.PencilGrids — Method

PencilGrids(length_tuple, resol_tuple::Tuple{Int, Int, Int})

Constructor for PencilGrids

Create an empty length[1]xlength[2]xlength[3] grid with resolution resol[1]xresol[2]xresol[3]. Each grid is a PencilArray, allowing multiprocess FFTs.

Examples

julia> using UltraDark

julia> PencilGrids((1.0, 1.0, 0.5), (64, 64, 32));

UltraDark.dV — Method

dV(grids)

Calculate the volume of each grid cell

Examples

julia> using UltraDark

julia> box_length = 1.0;

julia> resol = 16;

julia> g = Grids(box_length, resol);

julia> dV(g) * resol^3 == box_length^3
true

UltraDark.E_grav — Method

E_grav(grids)
E_grav(grids, psi)

Gravitational potential energy

UltraDark.E_kq — Method

E_kq(grids)
E_kq(grids, psi)

Sum of kinetic and quantum energies

UltraDark.actual_time_step — Method

actual_time_step(max_time_step, time_interval, time_step_options)

Actual size and number of time steps that should be taken if the maximum is max_time_step. No more than time_step_options.update_period steps should be taken, and they should fit in time_interval.

Examples

julia> using UltraDark: actual_time_step, TimeStepOptions

julia> actual_time_step(0.11, 1, TimeStepOptions())
(0.1, 10)

UltraDark.add_external_potential! — Method

add_external_potential!(t, grids, constants)

Add a gravitational potential to the grid. By default this does nothing, but can be overridden in multiple dispatch.

UltraDark.angular_momentum — Method

angular_momentum(grids)
angular_momentum(grids, ψx, ρx)

Calculate total angular momentum

Returns

L: AbstractArray with length 3

UltraDark.angular_momentum_density — Method

angular_momentum_density(grids)
angular_momentum_density(grids, ψx, ρx)

Calculate angular momentum density at each grid point

Returns

L: AbstractArray with dimensions 3 x resolx x resoly x resol_z

UltraDark.auxiliary_step! — Method

auxiliary_step!(Δt, grids, t, constants)

Do an auxiliary inner step. By default this does nothing, but can be overridden in multiple dispatch.

UltraDark.azimuthal_angle — Method

polar_angle(grids, r0)

Calculate the azimuthal angle in spherical or cylindrical coordinates

This is \phi in conventional physics notation.

UltraDark.evolve_to! — Method

Evolve grids forward from t_start to t_end

UltraDark.good_phase_diff — Method

bad_phase_diff(grids, config)

Check if the phase of the ψ field is in a trustable regime.

Returns

UltraDark.inner_step! — Method

inner_step!(Δt, grids, constants; a=1.0)

Perform the "inner" time step in the symmetrized split-step Fourier method.

This step applies the diffusion terms and updates the gravitational potential.

UltraDark.k_vec — Method

k_vec(lengths, resols)

Calculate the Fourier frequencies of a box with side lengths lengths and resolutions resols

Examples

julia> using UltraDark: k_vec

julia> kvec = k_vec((2π, 2π, 2π), (4, 4, 4));

julia> kvec[1]
4-element AbstractFFTs.Frequencies{Float64}:
  0.0
  1.0
 -2.0
 -1.0

UltraDark.mass — Method

total_mass(grids)
total_mass(grids, rho)

Calculate total mass of a density field

Examples

julia> using UltraDark

julia> g = Grids(1.0, 16);

julia> g.ρx .= 0.0;

julia> g.ρx[1, 1, 1] = 1.0;

julia> UltraDark.mass(g) == 1.0 * (1.0 / 16)^3
true

UltraDark.max_normed_phase_diff — Method

normed_max_phase_grad(grids)

Compute maximum phase gradient of a grid

Normalised to ignore large gradients in regions with low density. These tend to be anomalous.

UltraDark.max_time_step — Method

max_time_step(grids, a)

Calculate an upper bound on the time step

UltraDark.max_time_step — Method

max_time_step(grids, a)

Calculate an upper bound on the time step for PencilGrids

UltraDark.outer_step! — Method

outer_step!(Δt, grids, constants; a=1.0)

Perform the "outer" time step in the symmetrized split-step Fourier method.

This step only updates the phase of ψ applying accelerations due to gravity, the amplitude is not changed.

UltraDark.phase_diff — Method

phase_diff(field, dir)

Compute point-to-point difference of phase on a grid along a direction

Returns an array of size (size(field)[1], size(field)[2], size(field)[3]) containing differences in direction dir.

UltraDark.phase_diff — Method

phase_diff(field::PencilArray, dir)

Compute point-to-point difference of phase on a grid along a direction

Returns an array of size (size(field)[1], size(field)[2], size(field)[3]) containing differences in direction dir.

UltraDark.polar_angle — Method

polar_angle(grids, r0)

Calculate the polar angle in spherical coordinates

This is \theta in conventional physics notation.

UltraDark.radius_cylindrical — Method

radius_cylindrical(grids, r0)

Calculate the radial coordinate in cylindrical coordinates

Examples

julia> using UltraDark

julia> import UltraDark: radius_cylindrical, azimuthal_angle

julia> box_length = 1.0;

julia> resol = 16;

julia> g = Grids(box_length, resol);

julia> all(radius_cylindrical(g) .* cos.(azimuthal_angle(g)) .≈ g.x)
true

julia> all(
           radius_cylindrical(g, (0.0, 0.0, 1.0)) .* cos.(azimuthal_angle(g, (0.0, 0.0, 1.0))) .≈
           g.x,
       )
true

UltraDark.radius_spherical — Method

radius_spherical(grids, r0)

Calculate the radial coordinate in a spherical coordinate system

Examples

julia> using UltraDark

julia> import UltraDark: radius_spherical, polar_angle, azimuthal_angle

julia> box_length = 1.0;

julia> resol = 16;

julia> g = Grids(box_length, resol);

julia> all(radius_spherical(g) .* sin.(polar_angle(g)) .* cos.(azimuthal_angle(g)) .≈ g.x)
true

julia> all(
           radius_spherical(g, (1.0, 0.0, 0.0)) .* sin.(polar_angle(g, (1.0, 0.0, 0.0))) .*
           cos.(azimuthal_angle(g, (1.0, 0.0, 0.0))) .+ 1.0 .≈ g.x,
       )
true

UltraDark.take_steps! — Method

Take n steps with time step Δt

Examples

julia> using UltraDark: take_steps!, Grids, OutputConfig, Config

julia> take_steps!(
           Grids(1.0, 16),
           0,
           0.5,
           10,
           OutputConfig(mktempdir(), []),
           Config.constant_scale_factor,
           nothing,
       )
5.0