UltraDark.jl

UltraDark.Grids

UltraDark.Grids

UltraDark.Grids

UltraDark.PencilGrids

UltraDark.PencilGrids

UltraDark.PencilGrids

UltraDark.actual_time_step

UltraDark.add_external_potential!

UltraDark.evolve_to!

UltraDark.k_vec

UltraDark.max_normed_phase_grad

UltraDark.max_time_step

UltraDark.max_time_step

UltraDark.phase_diff

UltraDark.phase_diff

UltraDark.take_steps!

UltraDark.Grids — Type

struct containing grids used in a simulation

Examples

julia> using UltraDark

julia> len = 1;

julia> resol = 16;

julia> Grids(len, resol);

UltraDark.Grids — Method

Grids(length, resol::Int)

Constructor for Grids

Create an empty grid with length length and resolution resol

Examples

julia> using UltraDark

julia> Grids(1.0, 64);

UltraDark.Grids — Method

Grids(length_tuple, resol_tuple::Tuple{Int, Int, Int})

Constructor for Grids

Create an empty length[1]xlength[2]xlength[3] grid with resolution resol[1]xresol[2]xresol[3].

Examples

julia> using UltraDark

julia> Grids((1.0, 1.0, 0.5), (64, 64, 32));

UltraDark.PencilGrids — Type

struct containing grids used in a simulation

Examples

julia> using UltraDark

julia> len = 1;

julia> resol = 16;

julia> PencilGrids(len, resol);

UltraDark.PencilGrids — Method

PencilGrids(length, resol)

Constructor for PencilGrids

Create an empty grid with length length and resolution resol. Uses PencilFFTs to create PencilArrays.

Examples

julia> using UltraDark

julia> PencilGrids(1.0, 64);

UltraDark.PencilGrids — Method

PencilGrids(length_tuple, resol_tuple::Tuple{Int, Int, Int})

Constructor for PencilGrids

Create an empty length[1]xlength[2]xlength[3] grid with resolution resol[1]xresol[2]xresol[3]. Each grid is a PencilArray, allowing multiprocess FFTs.

Examples

julia> using UltraDark

julia> PencilGrids((1.0, 1.0, 0.5), (64, 64, 32));

UltraDark.actual_time_step — Method

actual_time_step(max_timestep, time_interval, n::Integer)

Actual size and number of time steps that should be taken if the maximum is max_timestep, no more than n steps should be taken, and they should fit in time_interval.

Examples

julia> using UltraDark: actual_time_step

julia> actual_time_step(0.11, 1, 20)
(0.1, 10)

UltraDark.add_external_potential! — Method

add_external_potential!(t, grids, constants)

Add a gravitational potential to the grid. By default this does nothing, but can be overridden in multiple dispatch.

UltraDark.evolve_to! — Method

Evolve grids forward from t_start to t_end

UltraDark.k_vec — Method

k_vec(lengths, resols)

Calculate the Fourier frequencies of a box with side lengths lengths and resolutions resols

Examples

julia> using UltraDark: k_vec

julia> kvec = k_vec((2π, 2π, 2π), (4, 4, 4));

julia> kvec[1]
4-element AbstractFFTs.Frequencies{Float64}:
  0.0
  1.0
 -2.0
 -1.0

UltraDark.max_normed_phase_grad — Method

normed_max_phase_grad(grids)

Compute maximum phase gradient of a grid

Normalised to ignore large gradients in regions with low density. These tend to be anomalous.

UltraDark.max_time_step — Method

max_time_step(grids, a)

Calculate an upper bound on the time step

UltraDark.max_time_step — Method

max_time_step(grids, a)

Calculate an upper bound on the time step for PencilGrids

UltraDark.phase_diff — Method

phase_diff(field, dir)

Compute point-to-point difference of phase on a grid along a direction

Returns an array of size (size(field)[1], size(field)[2], size(field)[2]) containing gradients in direction dir.

UltraDark.phase_diff — Method

phase_diff(field::PencilArray, dir)

Compute point-to-point difference of phase on a grid along a direction

Returns an array of size (size(field)[1], size(field)[2], size(field)[2]) containing gradients in direction dir.

UltraDark.take_steps! — Method

Take n steps with time step Δt

Examples

julia> using UltraDark: take_steps!, Grids, OutputConfig, Config

julia> take_steps!(Grids(1.0, 16), 0, 0.5, 10, OutputConfig(mktempdir(), []), Config.constant_scale_factor, nothing)
5.0