JultraDark.jl

JultraDark.Grids

JultraDark.Grids

JultraDark.Grids

JultraDark.PencilGrids

JultraDark.PencilGrids

JultraDark.PencilGrids

JultraDark.actual_time_step

JultraDark.evolve_to!

JultraDark.k_vec

JultraDark.max_normed_phase_grad

JultraDark.max_time_step

JultraDark.output_grids

JultraDark.output_summary_header

JultraDark.output_summary_row

JultraDark.phase_diff

JultraDark.take_steps!

JultraDark.Grids — Type

struct containing grids used in a simulation

Examples

julia> using JultraDark

julia> len = 1;

julia> resol = 16;

julia> Grids(len, resol);

JultraDark.Grids — Method

Grids(ψx::Array{Complex{Float64}}, length::Real)

Constructor for Grids

Create a grid with given ψ field, length length and resolution inferred from ψx

Examples

Can be contructed from a ψ field and box length,

julia> using JultraDark

julia> ψ = zeros(Complex{Float64}, 16, 16, 16);

julia> len = 1;

julia> Grids(ψ, len);

JultraDark.Grids — Method

Grids(length::Real, resol::Integer)

Constructor for Grids

Create an empty grid with length length and resolution resol

Examples

julia> using JultraDark

julia> Grids(1.0, 64);

JultraDark.PencilGrids — Type

struct containing grids used in a simulation

Examples

julia> using JultraDark

julia> len = 1;

julia> resol = 16;

julia> PencilGrids(len, resol);

JultraDark.PencilGrids — Method

PencilGrids(ψx::Array{Complex{Float64}}, length::Real)

Constructor for PencilGrids

Create a grid with given ψ field, length length and resolution inferred from ψx

Examples

Can be contructed from a ψ field and box length,

julia> using JultraDark

julia> ψ = zeros(Complex{Float64}, 16, 16, 16);

julia> len = 1;

julia> PencilGrids(ψ, len);

JultraDark.PencilGrids — Method

PencilGrids(length::Real, resol::Integer)

Constructor for PencilGrids

Create an empty grid with length length and resolution resol. Uses PencilFFTs to create PencilArrays.

Examples

julia> using JultraDark

julia> PencilGrids(1.0, 64);

JultraDark.actual_time_step — Method

actual_time_step(max_timestep::Real, time_interval::Real, n::Integer)

Actual size and number of time steps that should be taken if the maximum is max_timestep, no more than n steps should be taken, and they should fit in time_interval.

Examples

julia> using JultraDark: actual_time_step

julia> actual_time_step(0.11, 1, 20)
(0.1, 10)

JultraDark.evolve_to! — Method

Evolve grids forward from t_start to t_end

JultraDark.k_vec — Method

k_vec(lengths, resols)

Calculate the Fourier frequencies of a box with side lengths lengths and resolutions resols

Examples

julia> using JultraDark: k_vec

julia> kvec = k_vec((2π, 2π, 2π), (4, 4, 4));

julia> kvec[1]
4-element AbstractFFTs.Frequencies{Float64}:
  0.0
  1.0
 -2.0
 -1.0

JultraDark.max_normed_phase_grad — Method

normed_max_phase_grad(grids)

Compute maximum phase gradient of a grid

Normalised to ignore large gradients in regions with low density. These tend to be anomalous.

JultraDark.max_time_step — Method

max_time_step(grids, a)

Calculate an upper bound on the time step

JultraDark.output_grids — Method

output_grids(grids, output_config, step)

Write output from grids as specified in output_config

JultraDark.output_summary_header — Method

output_summary_header(output_config)

Write a header for a summary file

The header contains labels for each column of the summary CSV file. This function overwrites the current contents of the file.

JultraDark.output_summary_row — Method

output_summary_row(grids, output_config, t, a, Δt)

Write a new row to the summary file

JultraDark.phase_diff — Method

phase_diff(field, dir)

Compute point-to-point difference of phase on a grid along a direction

Returns an array of size (size(field)[1], size(field)[2], size(field)[2]) containing gradients in direction dir.

JultraDark.take_steps! — Method

Take n steps with time step Δt

Examples

julia> using JultraDark: take_steps!, Grids, OutputConfig

julia> take_steps!(Grids(1.0, 16), 0, 0.5, 10, OutputConfig(mktempdir(), []), t->1)
5.0